+++ src/actions/attack.cpp
@@ -183,13 +183,13 @@ battle_context_unit_stats::battle_context_unit_stats(const unit& u,
 			resources::gameboard->units(), resources::gameboard->map(), u_loc, u.alignment(), u.is_fearless());
 
 	// Leadership bonus.
-	int leader_bonus = under_leadership(units, u_loc).first;
+	int leader_bonus = under_leadership(units, u_loc, weapon).first;
 	if(leader_bonus != 0) {
 		damage_multiplier += leader_bonus;
 	}
 
 	// Resistance modifier.
-	damage_multiplier *= opp.damage_from(*weapon, !attacking, opp_loc);
+	damage_multiplier *= opp.damage_from(*weapon, !attacking, opp_loc, opp_weapon);
 
 	// Compute both the normal and slowed damage.
 	damage = round_damage(base_damage, damage_multiplier, 10000);
@@ -1579,14 +1579,14 @@ void attack_unit_and_advance(const map_location& attacker,
 	}
 }
 
-std::pair<int, map_location> under_leadership(const unit_map& units, const map_location& loc)
+std::pair<int, map_location> under_leadership(const unit_map& units, const map_location& loc, const_attack_ptr weapon)
 {
 	const unit_map::const_iterator un = units.find(loc);
 	if(un == units.end()) {
 		return {0, map_location::null_location()};
 	}
 
-	unit_ability_list abil = un->get_abilities("leadership");
+	unit_ability_list abil = un->get_abilities("leadership", weapon);
 	return abil.highest("value");
 }