+++ changelog.md
@@ -2,6 +2,9 @@
  ### WML engine
    * Support formula= key in [variable] ConditionalWML
    * Support to_location in [move_unit], taking a location ID
+   * Support [filter_weapon] in leadership and resistance abilities,
+     which activates the ability only when the affected unit is using
+     a matching weapon.
  ### Language and i18n
    * Fixed many cases of interpolated strings in the engine possibly having
      their translations retrieved from the wrong textdomain and falling back